prevent users from logging in more than once

+++ src/main/java/org/example/company/service/AuthService.java
@@ -7,7 +7,9 @@ import org.example.company.dto.register.RegisterRequestDTO;
 import org.example.company.dto.register.RegisterResponseDTO;
 import org.example.company.enums.UserRoles;
 import org.example.company.exception.UserAlreadyExistsException;
+import org.example.company.exception.UserAlreadyLoggedInException;
 import org.example.company.model.User;
+import org.example.company.repository.TokenRepository;
 import org.example.company.repository.UserRepository;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -28,6 +30,7 @@ public class AuthService {
     private final AuthenticationManager authenticationManager;
     private final JwtService jwtService;
     private final TokenService tokenService;
+    private final TokenRepository tokenRepository;
 
 
     public RegisterResponseDTO register(RegisterRequestDTO registerRequestDTO) throws UserAlreadyExistsException {
@@ -53,6 +56,12 @@ public class AuthService {
 
         User user = userRepository.findUserByUsername(loginRequestDTO.getUsername())
                 .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+        var validUserTokens = tokenService.getAllValidUserTokens(user);
+        if (!validUserTokens.isEmpty()) {
+           throw new UserAlreadyLoggedInException("You are already logged in");
+        }
+
         tokenService.deleteAllValidUserTokens(user);
         String jwtToken = jwtService.generateToken(user);
         tokenService.saveUserToken(user, jwtToken);
